+++ pennylane/capture/expand_transforms.py
@@ -21,7 +21,7 @@ from functools import wraps
 
 import jax
 
-from pennylane.transforms.core.transform import _create_transform_primitive
+from pennylane.core.transforms.transform import _create_transform_primitive  # tach-ignore
 
 from .base_interpreter import PlxprInterpreter
