+++ pennylane/capture/primitives.py
@@ -26,12 +26,12 @@ from pennylane.control_flow.for_loop import _get_for_loop_qfunc_prim
 from pennylane.control_flow.while_loop import _get_while_loop_qfunc_prim
 from pennylane.core._capture_measurements import _get_abstract_measurement  # tach-ignore
 from pennylane.core.operator.base import _get_abstract_operator  # tach-ignore
+from pennylane.core.transforms.transform import _create_transform_primitive  # tach-ignore
 from pennylane.ops.mid_measure.mid_measure import _create_mid_measure_primitive
 from pennylane.ops.mid_measure.pauli_measure import _create_pauli_measure_primitive
 from pennylane.ops.op_math.adjoint import _get_adjoint_qfunc_prim
 from pennylane.ops.op_math.condition import _get_cond_qfunc_prim
 from pennylane.ops.op_math.controlled import _get_ctrl_qfunc_prim
-from pennylane.transforms.core.transform import _create_transform_primitive
 from pennylane.workflow._capture_qnode import qnode_prim
 
 from .subroutine import quantum_subroutine_prim
