+++ pennylane/devices/default_clifford.py
@@ -27,6 +27,7 @@ import numpy as np
 from pennylane import math, ops
 from pennylane.core.operator import Channel, Operator, StatePrepBase
 from pennylane.core.qscript import QuantumScript, QuantumScriptBatch
+from pennylane.core.transforms import CompilePipeline
 from pennylane.exceptions import DeviceError, QuantumFunctionError
 from pennylane.measurements import (
     ClassicalShadowMP,
@@ -46,7 +47,6 @@ from pennylane.ops.qubit.observables import BasisStateProjector
 from pennylane.pauli import PauliWord, pauli_decompose
 from pennylane.pauli.utils import _binary_matrix_from_pws
 from pennylane.transforms import convert_to_numpy_parameters, defer_measurements, transform
-from pennylane.transforms.core import CompilePipeline
 from pennylane.typing import Result, ResultBatch
 
 from .default_qubit import accepted_sample_measurement
