+++ pennylane/devices/default_qubit.py
@@ -29,6 +29,7 @@ from pennylane import capture, math, ops
 from pennylane.core.measurements import MeasurementProcess, SampleMeasurement, StateMeasurement
 from pennylane.core.qscript import QuantumScript, QuantumScriptBatch, QuantumScriptOrBatch
 from pennylane.core.shots import Shots
+from pennylane.core.transforms import CompilePipeline
 from pennylane.decomposition.gate_set import GateSet
 from pennylane.exceptions import DecompositionUndefinedError, DeviceError
 from pennylane.logging import debug_logger, debug_logger_init
@@ -50,7 +51,7 @@ from pennylane.transforms import (
     defer_measurements,
     dynamic_one_shot,
 )
-from pennylane.transforms.core import CompilePipeline, transform
+from pennylane.transforms.core import transform
 from pennylane.typing import PostprocessingFn, Result, ResultBatch, TensorLike
 
 from .device_api import Device
