+++ pennylane/devices/default_qutrit_mixed.py
@@ -24,10 +24,10 @@ import numpy as np
 
 import pennylane as qp
 from pennylane.core.qscript import QuantumScript, QuantumScriptOrBatch
+from pennylane.core.transforms import CompilePipeline
 from pennylane.exceptions import DeviceError
 from pennylane.logging import debug_logger, debug_logger_init
 from pennylane.ops import _qutrit__channel__ops__ as channels
-from pennylane.transforms.core import CompilePipeline
 from pennylane.typing import Result, ResultBatch
 
 from . import Device
