+++ pennylane/devices/default_tensor.py
@@ -30,6 +30,7 @@ import pennylane as qp
 from pennylane.core.measurements import MeasurementProcess, StateMeasurement
 from pennylane.core.operator import Operation, Operator
 from pennylane.core.qscript import QuantumScript, QuantumScriptOrBatch
+from pennylane.core.transforms import CompilePipeline
 from pennylane.devices import Device, ExecutionConfig
 from pennylane.devices.modifiers import simulator_tracking, single_tape_support
 from pennylane.devices.preprocess import (
@@ -46,7 +47,6 @@ from pennylane.measurements import (
 )
 from pennylane.ops import LinearCombination, Prod, SProd, Sum
 from pennylane.templates.subroutines.time_evolution.trotter import _recursive_expression
-from pennylane.transforms.core import CompilePipeline
 from pennylane.typing import Result, ResultBatch, TensorLike
 
 has_quimb = True
