+++ pennylane/devices/legacy_facade.py
@@ -27,6 +27,7 @@ from pennylane import math, ops
 from pennylane.core.operator import Operator
 from pennylane.core.qscript import QuantumScript
 from pennylane.core.shots import Shots
+from pennylane.core.transforms import CompilePipeline, transform
 from pennylane.decomposition.gate_sets import ROTATIONS_PLUS_CNOT
 from pennylane.devices.capabilities import DeviceCapabilities
 from pennylane.exceptions import (
@@ -39,7 +40,6 @@ from pennylane.math import Interface, requires_grad
 from pennylane.measurements import ExpectationMP
 from pennylane.ops import MidMeasure
 from pennylane.transforms import broadcast_expand, defer_measurements, dynamic_one_shot
-from pennylane.transforms.core import CompilePipeline, transform
 from pennylane.wires import Wires
 
 from ._legacy_device import Device as LegacyDevice
