+++ pennylane/transforms/compile.py
@@ -19,8 +19,8 @@ from functools import partial
 import pennylane as qp
 from pennylane.core.qscript import QuantumScript, QuantumScriptBatch
 from pennylane.core.queuing import QueuingManager
+from pennylane.core.transforms import Transform, transform
 from pennylane.decomposition import gate_sets
-from pennylane.transforms.core import Transform, transform
 from pennylane.transforms.optimization import (
     cancel_inverses,
     commute_controlled,
