+++ pennylane/workflow/qnode.py
@@ -32,10 +32,10 @@ from pennylane import math, pytrees
 from pennylane.core.qscript import QuantumScript
 from pennylane.core.queuing import AnnotatedQueue
 from pennylane.core.shots import Shots, ShotsLike
+from pennylane.core.transforms import BoundTransform, CompilePipeline, Transform
 from pennylane.exceptions import QuantumFunctionError
 from pennylane.logging import debug_logger
 from pennylane.math import Interface
-from pennylane.transforms.core import BoundTransform, CompilePipeline, Transform
 from pennylane.typing import TensorLike
 
 from .execution import execute
