+++ tests/core/test_compile_pipeline.py
@@ -22,19 +22,19 @@ import rustworkx as rx
 
 import pennylane as qp
 from pennylane.core.qscript import QuantumScript, QuantumScriptBatch
+from pennylane.core.transforms.compile_pipeline import (
+    CotransformCache,
+    ProtectedLevel,
+    _apply_postprocessing_stack,
+    _batch_postprocessing,
+    null_postprocessing,
+)
 from pennylane.exceptions import QuantumFunctionError
 from pennylane.transforms.core import (
     BoundTransform,
     CompilePipeline,
     TransformError,
     transform,
-)
-from pennylane.transforms.core.compile_pipeline import (
-    CotransformCache,
-    ProtectedLevel,
-    _apply_postprocessing_stack,
-    _batch_postprocessing,
-    null_postprocessing,
 )
 from pennylane.typing import PostprocessingFn, Result, ResultBatch
