+++ tests/core/test_cotransform_cache.py
@@ -18,8 +18,8 @@ Tests for the CotransformCache object.
 import pytest
 
 import pennylane as qp
-from pennylane.transforms.core import BoundTransform
-from pennylane.transforms.core.cotransform_cache import CotransformCache
+from pennylane.core.transforms import BoundTransform
+from pennylane.core.transforms.cotransform_cache import CotransformCache
 
 
 def test_classical_jacobian_error_if_not_in_cache():
