+++ tests/devices/test_device_api.py
@@ -19,6 +19,7 @@ import pytest
 
 import pennylane as qp
 from pennylane.core.qscript import QuantumScript, QuantumScriptOrBatch
+from pennylane.core.transforms import CompilePipeline
 from pennylane.devices import Device, ExecutionConfig, MCMConfig
 from pennylane.devices.capabilities import (
     DeviceCapabilities,
@@ -26,7 +27,6 @@ from pennylane.devices.capabilities import (
     OperatorProperties,
 )
 from pennylane.exceptions import DeviceError, QuantumFunctionError
-from pennylane.transforms.core import CompilePipeline
 from pennylane.typing import Result, ResultBatch
 from pennylane.wires import Wires
