+++ tests/workflow/interfaces/run/test_run.py
@@ -20,8 +20,8 @@ import pytest
 import pennylane as qp
 from pennylane import numpy as pnp
 from pennylane.core.qscript import QuantumScript
+from pennylane.core.transforms import BoundTransform, CompilePipeline
 from pennylane.devices import ExecutionConfig
-from pennylane.transforms.core import BoundTransform, CompilePipeline
 from pennylane.transforms.optimization import merge_rotations
 from pennylane.workflow import run
